# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Future Jaano - AI-powered spiritual guidance platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMyProfile
      tags: [user]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMyProfile
      tags: [user]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /horoscope/daily/{sign}:
    get:
      operationId: getDailyHoroscopeBySign
      tags: [horoscope]
      summary: Get daily horoscope for a specific zodiac sign
      parameters:
        - name: sign
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Daily horoscope for the sign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HoroscopeEntry"
...
  /payments/initiate:
    post:
      operationId: initiatePayment
      tags: [payment]
      summary: Initiate a Razorpay payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrder"
  /payments/verify:
    post:
      operationId: verifyPayment
      tags: [payment]
      summary: Verify Razorpay payment signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVerification"
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResult"
...
  /push/subscribe:
    post:
      operationId: subscribePush
      tags: [notification]
      summary: Register a web push subscription for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSubscribeRequest"
      responses:
        "200":
          description: Subscription saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushSubscriptionStatus"
  /push/unsubscribe:
    post:
      operationId: unsubscribePush
      tags: [notification]
      summary: Remove a web push subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushUnsubscribeRequest"
      responses:
        "200":
          description: Subscription removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushSubscriptionStatus"
  /push/preferences:
    patch:
      operationId: updatePushPreferences
      tags: [notification]
      summary: Update notification time / language preferences for an existing subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushPreferencesUpdate"
      responses:
        "200":
          description: Preferences updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushSubscriptionStatus"
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, email, name, createdAt]
      properties:
        id: { type: integer }
        clerkId: { type: string }
        email: { type: string }
        name: { type: string }
        phone: { type: ["string", "null"] }
        dateOfBirth: { type: ["string", "null"] }
        timeOfBirth: { type: ["string", "null"] }
        placeOfBirth: { type: ["string", "null"] }
        gender: { type: ["string", "null"] }
        language: { type: string }
        isPremium: { type: boolean }
        avatarUrl: { type: ["string", "null"] }
        zodiacSign: { type: ["string", "null"] }
        createdAt: { type: string }
    UserProfileUpdate:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        dateOfBirth: { type: string }
        timeOfBirth: { type: string }
        placeOfBirth: { type: string }
        gender: { type: string }
        language: { type: string }
        zodiacSign: { type: string }
...
    HoroscopeEntry:
      type: object
      required: [sign, date, prediction, luckyColor, luckyNumber, compatibility]
      properties:
        sign: { type: string }
        signHindi: { type: ["string", "null"] }
        date: { type: string }
        prediction: { type: string }
        health: { type: string }
        career: { type: string }
        love: { type: string }
        finance: { type: string }
        luckyColor: { type: string }
        luckyNumber: { type: integer }
        luckyGem: { type: ["string", "null"] }
        compatibility: { type: string }
        rating: { type: integer }
        emoji: { type: string }

    # KUNDLI
...
    PaymentInput:
      type: object
      required: [planId, amount]
      properties:
        planId: { type: integer }
        amount: { type: integer }
        currency: { type: string }
    PaymentOrder:
      type: object
      required: [orderId, amount, currency, keyId]
      properties:
        orderId: { type: string }
        amount: { type: integer }
        currency: { type: string }
        keyId: { type: string }
        notes: { type: ["string", "null"] }
    PaymentVerification:
      type: object
      required: [razorpayOrderId, razorpayPaymentId, razorpaySignature, planId]
      properties:
        razorpayOrderId: { type: string }
        razorpayPaymentId: { type: string }
        razorpaySignature: { type: string }
        planId: { type: integer }
    PaymentResult:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean }
        message: { type: string }
        subscriptionId: { type: ["integer", "null"] }
...
    PushSubscribeRequest:
      type: object
      required: [endpoint, keys]
      properties:
        endpoint: { type: string }
        keys:
          type: object
          required: [p256dh, auth]
          properties:
            p256dh: { type: string }
            auth: { type: string }
        hour: { type: integer, minimum: 0, maximum: 23 }
        minute: { type: integer, minimum: 0, maximum: 59 }
        language: { type: string }
        userAgent: { type: string }
    PushUnsubscribeRequest:
      type: object
      required: [endpoint]
      properties:
        endpoint: { type: string }
    PushPreferencesUpdate:
      type: object
      required: [endpoint]
      properties:
        endpoint: { type: string }
        hour: { type: integer, minimum: 0, maximum: 23 }
        minute: { type: integer, minimum: 0, maximum: 59 }
        language: { type: string }
        enabled: { type: boolean }
    PushSubscriptionStatus:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
        count: { type: integer }
        message: { type: string }

    # PANCHANG
...